# openapi/market-bot.yaml
openapi: 3.1.0
info:
  title: Market Bot API
  version: 0.1.0
  description: Public API contract for the Market Bot backend.
servers:
  - url: /api/v1
paths:
...
  /carts/{cart_id}/items:
    post:
      operationId: addCartItem
      summary: Add an item to a cart
      parameters:
        - name: cart_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      example: USD
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
          format: uuid
...
    FulfillmentType:
      type: string
      enum:
        - digital
        - physical_standard
    CartItemSource:
      type: string
      enum:
        - user
        - ai
...
    CartItem:
      type: object
      required:
        - cart_item_id
        - product_id
        - variant_id
        - title
        - unit_price_minor
        - currency
        - quantity
        - source
        - fulfillment_type
        - available_stock
      properties:
        cart_item_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        title:
          type: string
        unit_price_minor:
          type: integer
          format: int64
          minimum: 0
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        quantity:
          type: integer
          format: int64
          minimum: 1
        source:
          $ref: '#/components/schemas/CartItemSource'
        fulfillment_type:
          $ref: '#/components/schemas/FulfillmentType'
        available_stock:
          type: integer
          format: int64
          minimum: 0
    Cart:
      type: object
      required:
        - cart_id
        - items
      properties:
        cart_id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
    AddCartItemRequest:
      type: object
      required:
        - product_id
      properties:
        product_id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int64
          minimum: 1
        source:
          $ref: '#/components/schemas/CartItemSource'
...
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
...
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      description: Admin token in the form admin/{user_id}
    userBearer:
      type: http
      scheme: bearer
      description: User token in the form user/{user_id}
  responses:
    InvalidInput:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'